Address reviewers feedback
